+++ app/build.gradle.kts
@@ -219,7 +219,7 @@ dependencies {
     implementation("org.jsoup:jsoup:1.15.3")
 
     // Job scheduling
-    implementation("androidx.work:work-runtime-ktx:2.6.0")
+    implementation("androidx.work:work-runtime-ktx:2.8.0")
     implementation("com.google.guava:guava:31.1-android")
 
     implementation("com.google.android.gms:play-services-gcm:17.0.0")
