+++ app/src/main/java/eu/kanade/tachiyomi/Migrations.kt
@@ -215,6 +215,8 @@ object Migrations {
                 if (!oldGroupHistory) {
                     preferences.groupChaptersHistory().set(RecentsPresenter.GroupType.Never)
                 }
+                LibraryUpdateJob.cancelAllWorks(context)
+                LibraryUpdateJob.setupTask(context)
             }
 
             return true
