+++ app/src/main/java/eu/kanade/tachiyomi/data/download/Downloader.kt
@@ -15,7 +15,7 @@ import eu.kanade.tachiyomi.data.database.models.Chapter
 import eu.kanade.tachiyomi.data.database.models.Manga
 import eu.kanade.tachiyomi.data.download.model.Download
 import eu.kanade.tachiyomi.data.download.model.DownloadQueue
-import eu.kanade.tachiyomi.data.library.LibraryUpdateService
+import eu.kanade.tachiyomi.data.library.LibraryUpdateJob
 import eu.kanade.tachiyomi.data.preference.PreferencesHelper
 import eu.kanade.tachiyomi.source.SourceManager
 import eu.kanade.tachiyomi.source.UnmeteredSource
@@ -304,7 +304,7 @@ class Downloader(
                     }
                 }
                 DownloadService.start(context)
-            } else if (!isRunning && !LibraryUpdateService.isRunning()) {
+            } else if (!isRunning && !LibraryUpdateJob.isRunning(context)) {
                 notifier.onDownloadPaused()
             }
         }
