+++ app/src/main/java/eu/kanade/tachiyomi/ui/recents/RecentsPresenter.kt
@@ -13,8 +13,7 @@ import eu.kanade.tachiyomi.data.download.DownloadService
 import eu.kanade.tachiyomi.data.download.DownloadServiceListener
 import eu.kanade.tachiyomi.data.download.model.Download
 import eu.kanade.tachiyomi.data.download.model.DownloadQueue
-import eu.kanade.tachiyomi.data.library.LibraryServiceListener
-import eu.kanade.tachiyomi.data.library.LibraryUpdateService
+import eu.kanade.tachiyomi.data.library.LibraryUpdateJob
 import eu.kanade.tachiyomi.data.preference.PreferencesHelper
 import eu.kanade.tachiyomi.source.SourceManager
 import eu.kanade.tachiyomi.ui.base.presenter.BaseCoroutinePresenter
@@ -49,7 +48,7 @@ class RecentsPresenter(
     val downloadManager: DownloadManager = Injekt.get(),
     val db: DatabaseHelper = Injekt.get(),
     private val chapterFilter: ChapterFilter = Injekt.get(),
-) : BaseCoroutinePresenter<RecentsController>(), DownloadQueue.DownloadListener, LibraryServiceListener, DownloadServiceListener {
+) : BaseCoroutinePresenter<RecentsController>(), DownloadQueue.DownloadListener, DownloadServiceListener {
 
     private var recentsJob: Job? = null
     var recentItems = listOf<RecentMangaItem>()
@@ -90,7 +89,7 @@ class RecentsPresenter(
         super.onCreate()
         downloadManager.addListener(this)
         DownloadService.addListener(this)
-        LibraryUpdateService.setListener(this)
+        LibraryUpdateJob.updateFlow.onEach(::onUpdateManga).launchIn(presenterScope)
         if (lastRecents != null) {
             if (recentItems.isEmpty()) {
                 recentItems = lastRecents ?: emptyList()
@@ -466,7 +465,6 @@ class RecentsPresenter(
     override fun onDestroy() {
         super.onDestroy()
         downloadManager.removeListener(this)
-        LibraryUpdateService.removeListener(this)
         DownloadService.removeListener(this)
         lastRecents = recentItems
     }
@@ -542,12 +540,12 @@ class RecentsPresenter(
         }
     }
 
-    override fun onUpdateManga(manga: Manga?) {
+    fun onUpdateManga(manga: Manga?) {
         when {
             manga == null -> {
                 presenterScope.launchUI { view?.setRefreshing(false) }
             }
-            manga.source == LibraryUpdateService.STARTING_UPDATE_SOURCE -> {
+            manga.source == LibraryUpdateJob.STARTING_UPDATE_SOURCE -> {
                 presenterScope.launchUI { view?.setRefreshing(true) }
             }
             else -> {
